+++ src/com/jec/commons/net/http/HttpMethod.ts
@@ -2,7 +2,7 @@
 //
 //   Copyright 2016-2017 Pascal ECHEMANN.
 //
-//   Licensed under the Apache License, Version 2.0 (the "License");
+//   Licensed under the Apache License, Version 2.0 (the "License"),
 //   you may not use this file except in compliance with the License.
 //   You may obtain a copy of the License at
 //
@@ -15,48 +15,47 @@
 //   limitations under the License.
 
 /**
- * The <code>HttpMethod</code> class provides an enumeration of static values 
- * that define HTTP methods.
+ * The <code>HttpMethod</code> Enum provides values that define HTTP methods.
  */
-export class HttpMethod {
+export enum HttpMethod {
 
   /**
    * Provides the reference for the <code>GET</code> HTTP method.
    */
-  public static readonly GET:string = "GET";
+  GET = "GET",
 
   /**
    * Provides the reference for the <code>POST</code> HTTP method.
    */
-  public static readonly POST:string = "POST";
+  POST = "POST",
 
   /**
    * Provides the reference for the <code>PUT</code> HTTP method.
    */
-  public static readonly PUT:string = "PUT";
+  PUT = "PUT",
 
   /**
    * Provides the reference for the <code>DELETE</code> HTTP method.
    */
-  public static readonly DELETE:string = "DELETE";
+  DELETE = "DELETE",
 
   /**
    * Provides the reference for the <code>HEAD</code> HTTP method.
    */
-  public static readonly HEAD:string = "HEAD";
+  HEAD = "HEAD",
 
   /**
    * Provides the reference for the <code>OPTIONS</code> HTTP method.
    */
-  public static readonly OPTIONS:string = "OPTIONS";
+  OPTIONS = "OPTIONS",
 
   /**
    * Provides the reference for the <code>TRACE</code> HTTP method.
    */
-  public static readonly TRACE:string = "TRACE";
+  TRACE = "TRACE",
 
   /**
    * Provides the reference for the <code>CONNECT</code> HTTP method.
    */
-  public static readonly CONNECT:string = "CONNECT";
+  CONNECT = "CONNECT"
 }